requirements.txt
openai>=1.0.0
gradio>=4.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
